requirements.txt
fastapi==0.115.0
pydantic==2.9.2
pydantic_core==2.23.4
uvicorn==0.31.0
python-dotenv==1.0.1
openai==1.50.0
